requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
Faker==4.0.0
Fiona==1.8.13
GDAL==2.4.2
geojson==2.5.0
geopandas==0.6.2
idna==2.8
munch==2.5.0
mysql-connector-python==8.0.17
nose==1.3.7
numpy==1.16.4
pandas==0.25.3
protobuf==3.9.2
PyCristoforo==2.0.0
pyproj==2.4.2.post1
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
requests==2.22.0
Shapely==1.6.4.post2
six==1.12.0
text-unidecode==1.3
urllib3==1.25.6
